requirements.txt
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.30.0
azure-storage-blob==12.16.0
azure-identity==1.13.0